+++ src/main/java/keeper/project/homepage/controller/ExceptionController.java
@@ -1,6 +1,6 @@
 package keeper.project.homepage.controller;
 
-import keeper.project.homepage.exception.CustomAuthenticationEntryPointException;
+import keeper.project.homepage.exception.sign.CustomAuthenticationEntryPointException;
 import keeper.project.homepage.dto.result.CommonResult;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.access.AccessDeniedException;
