+++ src/main/java/keeper/project/homepage/exception/sign/CustomAuthenticationEntryPointException.java
@@ -1,4 +1,4 @@
-package keeper.project.homepage.exception;
+package keeper.project.homepage.exception.sign;
 
 public class CustomAuthenticationEntryPointException extends RuntimeException {
