+++ src/main/java/keeper/project/homepage/exception/sign/CustomSignUpFailedException.java
@@ -1,4 +1,4 @@
-package keeper.project.homepage.exception;
+package keeper.project.homepage.exception.sign;
 
 public class CustomSignUpFailedException extends RuntimeException {
